[#82] - Plan members slice & saga
